% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{myf}
\alias{myf}
\title{Squared function}
\usage{
myf(x)
}
\arguments{
\item{x}{}
}
\value{
a vetor of squared components
}
\description{
Takes a vector and returns the vetor with its components squared
}
\examples{
x=1:30; myf(x)
}
